# openapi-specification.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/grades/{studentId}":{"put":{"tags":["grade-controller"],"operationId":"updateGrades","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GradeBookItem"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GradeBookItem"}}}}}}},"/api/tasks":{"get":{"tags":["task-controller"],"operationId":"getTasks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskItem"}}}}}}},"post":{"tags":["task-controller"],"operationId":"createTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskItem"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskItem"}}}}}}},"/api/chats/{chatId}/messages":{"get":{"tags":["chat-controller"],"operationId":"getMessages","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageItem"}}}}}}},"post":{"tags":["chat-controller"],"operationId":"sendMessage","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageItem"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageItem"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/attendance/{studentId}":{"post":{"tags":["attendance-controller"],"operationId":"updateTodayStatus","parameters":[{"name":"studentId","in":"path","required":true,"schema":{"type":"string"}},{"name":"status","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AttendanceRegistryItem"}}}}}}},"/api/students":{"get":{"tags":["student-controller"],"operationId":"getStudents","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StudentItem"}}}}}}}},"/api/grades":{"get":{"tags":["grade-controller"],"operationId":"getGrades","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GradeBookItem"}}}}}}}},"/api/dashboard/stats":{"get":{"tags":["dashboard-controller"],"operationId":"getStats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StatItem"}}}}}}}},"/api/dashboard/alerts":{"get":{"tags":["dashboard-controller"],"operationId":"getAlerts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AcademicAlertItem"}}}}}}}},"/api/courses":{"get":{"tags":["course-controller"],"operationId":"getCourses","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CourseItem"}}}}}}}},"/api/chats":{"get":{"tags":["chat-controller"],"operationId":"getChats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatRoom"}}}}}}}},"/api/attendance":{"get":{"tags":["attendance-controller"],"operationId":"getAttendance","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AttendanceRegistryItem"}}}}}}}}},"components":{"schemas":{"GradeBookItem":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"code":{"type":"string"},"course":{"type":"string"},"group":{"type":"string"},"pc1":{"type":"number","format":"double"},"pc2":{"type":"number","format":"double"},"parcial":{"type":"number","format":"double"},"finalExam":{"type":"number","format":"double"},"average":{"type":"number","format":"double"},"status":{"type":"string"}}},"TaskItem":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"course":{"type":"string"},"group":{"type":"string"},"publishedDate":{"type":"string"},"limitDate":{"type":"string"},"receivedCount":{"type":"integer","format":"int32"},"totalCount":{"type":"integer","format":"int32"},"status":{"type":"string"}}},"MessageItem":{"type":"object","properties":{"id":{"type":"string"},"senderId":{"type":"string"},"text":{"type":"string"},"timestamp":{"type":"string"},"sentByMe":{"type":"boolean"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"email":{"type":"string"},"name":{"type":"string"},"role":{"type":"string"}}},"AttendanceRegistryItem":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"code":{"type":"string"},"course":{"type":"string"},"group":{"type":"string"},"attendance":{"type":"integer","format":"int32"},"todayStatus":{"type":"string"}}},"StudentItem":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"code":{"type":"string"},"course":{"type":"string"},"group":{"type":"string"},"averageGrade":{"type":"number","format":"double"},"attendance":{"type":"integer","format":"int32"},"status":{"type":"string"}}},"StatItem":{"type":"object","properties":{"label":{"type":"string"},"value":{"type":"object"},"subtext":{"type":"string"},"iconName":{"type":"string"}}},"AcademicAlertItem":{"type":"object","properties":{"id":{"type":"string"},"type":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"actionLabel":{"type":"string"}}},"CourseItem":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"code":{"type":"string"},"group":{"type":"string"},"studentsCount":{"type":"integer","format":"int32"},"averageGrade":{"type":"string"},"progress":{"type":"integer","format":"int32"},"nextClass":{"type":"string"},"status":{"type":"string"}}},"ChatContact":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"avatar":{"type":"string"},"role":{"type":"string"},"type":{"type":"string"},"online":{"type":"boolean"}}},"ChatRoom":{"type":"object","properties":{"id":{"type":"string"},"contact":{"$ref":"#/components/schemas/ChatContact"},"unreadCount":{"type":"integer","format":"int32"},"messages":{"type":"array","items":{"$ref":"#/components/schemas/MessageItem"}}}}}}}